refactor code and add contact address logic
