+++ include/address_book/address_book.py
@@ -1,0 +1,17 @@
+from collections import UserDict
+from include.upcoming_birthdays.upcoming_birthdays import get_upcoming_birthdays
+
+
+class AddressBook(UserDict):
+    def add_record(self, record):
+        self.data[record.name.value] = record
+
+    def find(self, name):
+        return self.data.get(name)
+
+    def delete(self, name):
+        if name in self.data:
+            del self.data[name]
+
+    def upcoming_birthdays(self):
+        return get_upcoming_birthdays(self)
